Implement landing page
Add a first design for the index page, based on the drafts.

It contains two a base that can be copied over for all other files that
defines the navbar. It's main components are two large buttons that
refere to the subsites as well as an About section (currently
placeholder).