Redo register.html
- include Font Awesome
- introduce divs for both sections
- redo registration form completly
- replace images with css and buttons

Both 'sections' of the page were in one container.
The registration form was not following bootstrap guidelines, labels
were missing or were not named. Inputs did not use a specific width for
the grid system and so on.
The oauth images also were not styled at all and just sat there.

This commit fixes the form and replaces it with what bootstrap actually
suggests. It also wraps the content into two divs so they can be
distinguished in the markup.
Additionally the unstyled images have been replaced with styled links.

This has lots of advantages. The form now checks for input
automatically. The site looks better on mobile. Screenreaders and other
accessablity tools actually manage to parse the form.