Implement 'add.html'
add.html represents the interface used to add lost elements to the
database. It features a form, containing a description text field, a
photo section that is fillable via a button/file selector and drag/drop
events as well as an event selection.

The demo component of the form loads selected files in the preview area
via base64 encoding.